Rename podlogsprocessor to containerprocessor
