+++ deploy/helm/events-collector-config.yaml
@@ -43,7 +43,7 @@ extensions:
     endpoint: 0.0.0.0:13133
 
 processors:
-  podlogsprocessor:
+  containerprocessor:
   memory_limiter:
 {{ toYaml .Values.otel.events.memory_limiter | indent 4 }}
 
@@ -379,9 +379,9 @@ service:
         - transform/set_status_for_pods
         - transform/stringify_body
         - resource/manifest
-#        - resourcedetection/providers
+#        - resourcedetection/providers TODO: Remove
         - swk8sattributes
-        - podlogsprocessor
+        - containerprocessor
         - transform/cleanup_attributes_for_nonexisting_entities
 {{- if not (and .Values.otel.events.enabled .Values.otel.manifests.enabled) }}
         - filter/k8s_collector_config_include
